[FIX] Fix login logout button
