+++ src/components/Header/Header.jsx
@@ -4,6 +4,8 @@ import { NavLink, Link } from "react-router-dom";
 import { ToggleButton } from "../../utils/ToggleButton";
 import Cookies from "js-cookie";
 import { useTranslation } from "react-i18next";
+import {useHistory} from "react-router-dom"
+import {ArrowArcLeft} from "phosphor-react"
 
 function Header(props) {
 
@@ -12,6 +14,15 @@ function Header(props) {
   const [cookie, setCookie] = useState(null);
   const [adminCookie, setAdminCookie] = useState(null);
   const [top, setTop] = useState(true);
+  const navigate = useHistory();
+
+  const logoutUser = async () => {
+    Cookies.remove("acess_token");
+    Cookies.remove("admin_token")
+    navigate.push("/");
+    navigate.go(0)
+  };
+
   useEffect(() => {
     let userToken = Cookies.get("acess_token");
     setCookie(userToken);
@@ -74,6 +85,13 @@ function Header(props) {
                 <NavLink exact to="/wishlist" activeClassName={styles.active}>
                 {t('Favoritos')}
                 </NavLink>
+              </li>
+              <li>
+              <button onClick={() => logoutUser()}>
+          <div className="z-10 lg:flex lg:items-center lg:gap-4 lg:fixed lg:bottom-12 lg:right-12 lg:rounded-lg lg:dark:bg-indigo-900 bg-blue-700 lg:hover:scale-105 lg:p-4 text-white transition-all duration-200 ease-in-out hidden">
+            <ArrowArcLeft></ArrowArcLeft>{t('Logout')}
+          </div>
+        </button>
               </li>
               </>
           ) : (
